🔨 modularize data catalog into files
