+++ site/DataCatalog/DataCatalog.scss
@@ -1,4 +1,4 @@
-#charts-index-page-root {
+#data-catalog-page-root {
     min-height: calc(100vh - $header-height-md);
     @include sm-only {
         min-height: calc(100vh - $header-height-sm);
@@ -218,6 +218,7 @@
         margin-left: 8px;
     }
 }
+
 .data-catalog-country-selector-switch {
     label {
         white-space: wrap;
@@ -313,15 +314,6 @@
     &:hover {
         background-color: $accent-pale-blue;
     }
-}
-
-.data-catalog-search-reset-button {
-    display: none;
-}
-
-/* Results */
-.data-catalog-facets {
-    margin-bottom: 16px;
 }
 
 .data-catalog-refinement-list {
