+++ site/owid.scss
@@ -39,7 +39,7 @@
 @import "css/noscript.scss";
 @import "./collections/CollectionsPage.scss";
 @import "./SiteSearchNavigation.scss";
-@import "./DataCatalog.scss";
+@import "./DataCatalog/DataCatalog.scss";
 @import "./site/search/Autocomplete.scss";
 @import "./site/search/ChartHit.scss";
 @import "./site/search/Search.scss";
